Fix graph plotting to handle year boundaries
Prior to this change, gnuplot was treating the week ID (example "201742"
as a number). This meant that any data that crossed the boundary between
years was plotted weird, because gnuplot thought there was missing data.
Example: data would "jump" from "201752" to "201801" (roughly 50
"missing" rows).

Solution is to plot data sequentially, not using the week ID as part of
the plot. I.e.: instead of "using 1:44" do "using 44". This however,
meant that the labels on the x-axis were no longer present. The solution
to this is to use xticlabels or "xtic" for short. The problem with this,
however, was that it was showing a label for every single tic mark (too
many).

The solution was to create a function which returns the labels every N
number of them, where N is automatically computed using two pieces of
information: the number of rows (records) in the data and the desired
number of tic marks.