Add option to only process weeks in a time window
Add options --after and --before to allow the user to specify a time
range to analyze. So instead of simply analyzing all of the data
included in a file(s), user can now choose to analyze data after a
particular point in time and/or before a particular point in time.